+++ src/chanki/week1/BJ10866.java
@@ -10,7 +10,7 @@ import java.util.Deque;
 import java.util.LinkedList;
 import java.util.StringTokenizer;
 
-public class B10866 {
+public class BJ10866 {
     public static void main(String[] args) throws IOException {
         String command, str;
         int cnt;
