+++ src/chanki/week1/BJ1406.java
@@ -10,7 +10,7 @@ import java.util.ArrayList;
 import java.util.Stack;
 import java.util.StringTokenizer;
 
-public class B1406 {
+public class BJ1406 {
     public static void main(String[] args) throws IOException {
         int command_num;
         String input, command;
