+++ src/chanki/week2/BJ10872.java
@@ -1,4 +1,4 @@
-// 팩토리얼 0의 갯수
+// 팩토리얼
 package chanki.week2;
 
 import java.io.BufferedReader;
@@ -6,10 +6,9 @@ import java.io.BufferedWriter;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 
-public class B1676 {
+public class BJ10872 {
     public static void main(String[] args) throws Exception{
-        int num, temp;
-        int cnt = 0;
+        int num;
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
@@ -17,19 +16,11 @@ public class B1676 {
         if(num == 0)
             num = 1;
 
-        for(int i = 1; i <= num; i++) {
-            temp = i;
-            while(true) {
-                if(temp % 5 == 0) {
-                    temp /= 5;
-                    cnt++;
-                } else {
-                    break;
-                }
-            }
+        for(int i = (num - 1); i > 1; i--) {
+            num *= i;
         }
 
-        bw.write(cnt + "");
+        bw.write(num + "");
 
         bw.flush();
         bw.close();
