+++ src/chanki/week2/BJ1676.java
@@ -1,4 +1,4 @@
-// 팩토리얼
+// 팩토리얼 0의 갯수
 package chanki.week2;
 
 import java.io.BufferedReader;
@@ -6,9 +6,10 @@ import java.io.BufferedWriter;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 
-public class B10872 {
+public class BJ1676 {
     public static void main(String[] args) throws Exception{
-        int num;
+        int num, temp;
+        int cnt = 0;
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
@@ -16,11 +17,19 @@ public class B10872 {
         if(num == 0)
             num = 1;
 
-        for(int i = (num - 1); i > 1; i--) {
-            num *= i;
+        for(int i = 1; i <= num; i++) {
+            temp = i;
+            while(true) {
+                if(temp % 5 == 0) {
+                    temp /= 5;
+                    cnt++;
+                } else {
+                    break;
+                }
+            }
         }
 
-        bw.write(num + "");
+        bw.write(cnt + "");
 
         bw.flush();
         bw.close();
